Panels, Horizontal and Left Column
I wrote the code for Panels and input it into the ui elements, and i took the liberty of changing some code(sort of) so it was easier to read in horizontal and left column cshtml files. I feel like this would make it easier to read if something needed to be changed. If you like it let me know and I will gladly put it into all the files.